+++ src/Three20Style/Classes/TTStyleSheet.m
@@ -108,7 +108,7 @@ static TTStyleSheet* gStyleSheet = nil;
 - (TTStyle*)styleWithSelector:(NSString*)selector forState:(UIControlState)state {
   NSString* key = state == UIControlStateNormal
     ? selector
-    : [NSString stringWithFormat:@"%@%d", selector, state];
+    : [NSString stringWithFormat:@"%@%u", selector, (unsigned int)state];
   TTStyle* style = [_styles objectForKey:key];
   if (!style) {
     SEL sel = NSSelectorFromString(selector);
